Se agrega ejemplo de navigacion entre pantallas
