+++ lib/src/presentation/pages/first_page.dart
@@ -1,0 +1,19 @@
+import 'package:flutter/material.dart';
+
+class FirstPage extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: Text('First Page'),
+      ),
+      body: Center(
+        child: Text('First'),
+      ),
+      floatingActionButton: FloatingActionButton(
+        child: Icon(Icons.arrow_forward_rounded),
+        onPressed: () => Navigator.pushNamed(context, '/second'),
+      ),
+    );
+  }
+}
