+++ lib/src/presentation/pages/pages.dart
@@ -1,0 +1,3 @@
+export 'home_page.dart';
+export 'first_page.dart';
+export 'second_page.dart';
