+++ lib/src/presentation/pages/second_page.dart
@@ -1,0 +1,20 @@
+import 'package:flutter/material.dart';
+
+class SecondPage extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: Text('Second Page'),
+      ),
+      body: Center(
+        child: Text('Second'),
+      ),
+      floatingActionButton: FloatingActionButton(
+        child: Icon(Icons.arrow_back),
+        backgroundColor: Colors.red,
+        onPressed: () => Navigator.pop(context),
+      ),
+    );
+  }
+}
